Implement graceful shutdown of HTTP server
When the of-watchdog receives a SIGTERM, it should stop accepting
new requests. The current processes in progress should continue up to
the write_timeout property value.

A new environment variable was also ported over from the original faas
watchdog 'suppress_lock' that will not write a lock file to disk,
and not provide Swarm with healthchecks.

This change is largely copied from the original faas watchdog
Tested with the 'hey' tool by invoking functions and forcing the scale
to 1 while monitoring the logs to verify the in-flight requests were
completed before shutting down the container

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>