Add http_buffer to "http" mode
Allows servers such as Swoole which do not implement the whole
HTTP spec to be used with of-watchdog. Apply this setting when
Transfer-Encoding: chunked is not compatible with your http
server in http mode.

Tested with netcat "nc" outside of a Docker container to see
the output either sent with a content-length when turned on
or without one when in chunked mode.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>