Add custom metrics
Custom metrics are added for HTTP traffic and exposed by default
on port 8081. This has been tested more extensively in the
classic watchdog and copied / duplicated into the of-watchdog.

Local e2e testing was done on MacOS outside of a container and
the new http_* fields and metrics were reported as expected.

Signed-off-by: Alex Ellis <alexellis2@gmail.com>