Fix feedback from #87
This commit applies the feedback given in #87 and uses a flag
to both add documentation and to match the style of the rest
of the project.

This change is only useful for exec healthchecks and not for
http probes. Wherever possible users should only use http
healthchecks since exec is very resource intensive.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>