Add mechanism to disable prefixes to function logs
This patch allows users to disable any prefix of date/time and
the text such as stderr or stdout which showed users what kind
of stream the log statement came from.

This was tested in HTTP mode with the watchdog running outside
of a container with Node.js printing two statements - one from
stderr and one from stdout. As expected, when the new
flag was toggled, neither had a prefix.

Fixes: #75 #82 #72

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>