Perform graceful shutdown according to number of connections
Prior to this change the watchdog would attempt to wait for
the write_timeout value twice, even if there were no connections
in flight.

This adds a new metric for in-flight connections and removes
the explicit wait.

Testing with killall of-watchdog -s SIGTERM performs as
expected, and clients that already connected complete their
work before the termination.

This change is a prerequisite for a change in faas-netes
https://github.com/openfaas/faas-netes/issues/853

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>