Allow for very long log lines
With this environment-variable (log_buffer_size), very large
logs can be scanned and printed out to stdio.

This was reported by RateHub (an OpenFaaS Pro) user.

Tested with a 5MB JSON file and a Node.js process that
printed the value to console.log/console.error.

It worked as expected. After reducing the value to the
default, an error was returned.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>