Fix issue with streaming mode and killing long running processes
Fixes an issue with SEGFAULTs or leaked child processes when running in
streaming mode.

Previously the watchdog implemented custom logic based on a timer to kill function
processes that exceed some `exec_timeout`. This solution was susceptible to
panics and resource leaks:

 - SIGSEGV when attempting to invoke `Kill()` on an invalid `*Cmd` pointer
 - OOMKills in Kubernetes due to what appears to be the watchdog failing to reap
child processes before executing new function call invocations.

With this change the entire timer based solution is discarded in favour
`CommandContext` from the `exec` package.

Tested with hey and a long running process. The process is still killed as
expected, but will need further verifcation from: @cmacq2 due to a lack
of instructions to reproduce the described issue.

Fixes: #138
Closes: #139

Signed-off-by: Johan Ouwerkerk <jm.ouwerkerk@gmail.com>
Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>