Fix: replace custom timer logic with standard context deadlines
Gets rid of segfaults and leaked child processes when running in streaming mode.

Previously the OpenFAAS watchdog implemented custom logic based on a timer to kill function processes that exceed some `exec_timeout`.
This solution was susceptible to panics and resource leaks:

 - SIGSEGV when attempting to invoke `Kill()` on an invalid `*Cmd` pointer
 - OOMKills in Kubernetes due to what appears to be the watchdog failing to reap child processes before executing new function call invocations.

With this change the entire timer based solution is discarded in favour of standard `CommandContext` API from the `exec` package.

Fixes: #138
Signed-off-by: Johan Ouwerkerk <jm.ouwerkerk@gmail.com>