feat: allow proxying readiness checks to the function
Allow setting an endpoint path for the function readiness check via an
ENV variable `function_ready_endpoint`

When this value is set, the requests to `/_/ready` will execute an empty
GET request with/to the configured endpoint. This allows the function
authors to implement custom readiness check logic.

This custom request is checked _after_ the the standard liviness checks
and the ConcurrencyLimiter check. For a completely custom readiness
check, the function should be deployed with `max_inflight == 0` and
`function_ready_endpoint` to the custom path.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>