feat: implement auth result parsing
Support parsing structured auth policy results. This allows the policy
to customize the error response code. The policy can also attach
additional headers to the request so that the implementation. The
primary use case for the additional headers is to pass user or client
identification to the implmentation to allow for auditing and/or
additional authorization checks.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>