Stream stderr to logs for streaming fork mode
This mode is not expected to be used by any customers in
production, however we received a message from a customer who
had decided to use this mode.

They reported not being able to log to stderr, I was able to
reproduce the problem and have added bindLoggingPipe to
stderr as per the other modes in the of-watchdog.

After testing with a simple Python program, the logging is
now available.

We only recommend HTTP mode for production use.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>