Provide option to timeout functions earlier than configured
A header of X-Timeout can be passed to reduce the value of
exec_timeout. The consumer invoking the function is responsible
for tuning this value and making sure it has enough headroom.

The change was requested by @kevin-lindsay-1 for functions
which process batched data of varying size and could run for
24 hours or 2 minutes, but need to timeout in a timely
fashion, according to the input.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>