Copy upstream response into http writer
As an optimization, the response body for the HTTP runner can
stream data back, instead of buffering it all in memory and
then sending it in one shot.

This change means replacing an io.ReadAll with an io.Copy, as a
result, it's possible that some callers will no longer receive
a content-length, but a stream of data with a length of -1.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>