Add "inproc" mode for Golang functions
The watchdog can be imported by a Go template in order to
avoid any hop over HTTP.

On the plus side - RAM looks to be much smaller as low as
2mb total usage vs the two processes.

On the downside, it's not possible or practical to do
stderr/stdout prefixing due to no process boundary between
the watchdog and the handler process.

A new template and Dockerfile will be required to support
the changes. Builds may be marginally slower since they have
to import the watchdog as an "SDK".

Every push will be a single layer binary so this will also
take longer vs a tiny Go binary plus the watchdog.

Initial testing in Kubernetes with a modified template
appears to work as expected.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>